+++ examples/cdn/outputs.tf
@@ -1,0 +1,3 @@
+output "domain" {
+  value = "${alicloud_cdn_domain.domain.id}"
+}